Added initial watchOS dumping
Needs work on GDCWebServer in the iOS companion application, and also need to ensure our dumping is more reliable - putting it in our interface controller is not a smart move.

Changes were made to RTBRuntime to faciliate loading and subsequently unloading frameworks individually. This is to avoid the memory usage limit imposed by watchOS on applications, as loading all frameworks would exceed that.

Changes were also made to RTBClass to access the resultant header as an NSString without writing it to disk first.